# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/correction-reports": {
      "get": {
        "operationId": "adminListCorrectionReports",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "new",
                "in_review",
                "resolved",
                "dismissed"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CorrectionReportQueueRowDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/correction-reports/{id}": {
      "get": {
        "operationId": "adminGetCorrectionReportPrivateDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrectionReportPrivatePayloadDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/correction-reports/{id}/claim": {
      "post": {
        "operationId": "adminClaimCorrectionReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimCorrectionReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrectionReportPrivateDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/correction-reports/{id}/correction": {
      "post": {
        "operationId": "adminLinkCorrectionReportEdit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCorrectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingCorrectionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/correction-reports/{id}/disposition": {
      "post": {
        "operationId": "adminDisposeCorrectionReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisposeCorrectionReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrectionReportPrivateDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/correction-reports/reporters/{reporterId}/abuse-restriction": {
      "post": {
        "operationId": "adminConfirmCorrectionReporterRestriction",
        "parameters": [
          {
            "name": "reporterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestrictCorrectionReporterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "family-events-api",
    "description": "Backend API for the Family Events platform. This OpenAPI document is the contract; the web app consumes a client generated from it.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "clerk": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CorrectionReportQueueRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "event_id": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "type": "string",
            "enum": [
              "cancellation",
              "wrong_date_time",
              "wrong_location",
              "wrong_cost",
              "accessibility",
              "other"
            ]
          },
          "priority": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "in_review",
              "resolved",
              "dismissed"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "claimed_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "claimed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "resolved_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "resolved_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "correction_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "event_title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "event_id",
          "category",
          "priority",
          "status",
          "version",
          "claimed_by",
          "claimed_at",
          "resolved_by",
          "resolved_at",
          "correction_id",
          "created_at",
          "updated_at",
          "event_title"
        ]
      },
      "CorrectionReportContactDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320
          },
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 32,
            "pattern": "^\\+?[0-9 ()-]{7,32}$"
          }
        }
      },
      "CorrectionReportPrivatePayloadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "event_id": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "type": "string",
            "enum": [
              "cancellation",
              "wrong_date_time",
              "wrong_location",
              "wrong_cost",
              "accessibility",
              "other"
            ]
          },
          "priority": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "in_review",
              "resolved",
              "dismissed"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "claimed_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "claimed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "resolved_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "resolved_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "correction_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "reporter_user_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "details": {
            "type": "string"
          },
          "resolution_note": {
            "type": "string",
            "nullable": true
          },
          "event_title": {
            "type": "string"
          },
          "contact": {
            "nullable": true,
            "description": "Private optional reporter contact",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CorrectionReportContactDto"
              }
            ]
          },
          "evidence": {
            "nullable": true,
            "description": "Private supporting HTTP(S) URLs",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "event_id",
          "category",
          "priority",
          "status",
          "version",
          "claimed_by",
          "claimed_at",
          "resolved_by",
          "resolved_at",
          "correction_id",
          "created_at",
          "updated_at",
          "reporter_user_id",
          "details",
          "resolution_note",
          "event_title",
          "contact",
          "evidence"
        ]
      },
      "ClaimCorrectionReportDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "version"
        ]
      },
      "CorrectionReportPrivateDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "event_id": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "type": "string",
            "enum": [
              "cancellation",
              "wrong_date_time",
              "wrong_location",
              "wrong_cost",
              "accessibility",
              "other"
            ]
          },
          "priority": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "in_review",
              "resolved",
              "dismissed"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "claimed_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "claimed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "resolved_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "resolved_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "correction_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "reporter_user_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "details": {
            "type": "string"
          },
          "resolution_note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "event_id",
          "category",
          "priority",
          "status",
          "version",
          "claimed_by",
          "claimed_at",
          "resolved_by",
          "resolved_at",
          "correction_id",
          "created_at",
          "updated_at",
          "reporter_user_id",
          "details",
          "resolution_note"
        ]
      },
      "LinkCorrectionDto": {
        "type": "object",
        "properties": {
          "audit_log_id": {
            "type": "string",
            "format": "uuid"
          },
          "note": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "audit_log_id",
          "note"
        ]
      },
      "ListingCorrectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "event_id": {
            "type": "string",
            "format": "uuid"
          },
          "operator_id": {
            "type": "string",
            "format": "uuid"
          },
          "audit_log_id": {
            "type": "string",
            "format": "uuid"
          },
          "audit_note": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "event_id",
          "operator_id",
          "audit_log_id",
          "audit_note",
          "created_at"
        ]
      },
      "DisposeCorrectionReportDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "outcome": {
            "type": "string",
            "enum": [
              "resolved",
              "dismissed"
            ]
          },
          "note": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "correction_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "version",
          "outcome",
          "note"
        ]
      },
      "RestrictCorrectionReporterDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "reason",
          "expires_at"
        ]
      },
      "OkDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
...
    }
  }
}